Restructure test suite for targeted execution
