+++ tests/data/batteryAttributes.test.js
@@ -1,4 +1,4 @@
-const devices = require('../data');
+const devices = require('../../data');
 
 test('batteries expose pinA attribute', () => {
   const batteryGroups = [devices.batteries || {}, (devices.accessories && devices.accessories.batteries) || {}];
