+++ tests/data/checkConsistency.test.js
@@ -1,4 +1,4 @@
-const checkConsistency = require('../checkConsistency');
+const checkConsistency = require('../../checkConsistency');
 
 test('camera data includes required fields', () => {
   const result = checkConsistency();
