+++ tests/data/generateSchema.test.js
@@ -1,4 +1,4 @@
-const { buildSchema, isDeviceObject, isDeviceMap } = require('../generateSchema');
+const { buildSchema, isDeviceObject, isDeviceMap } = require('../../generateSchema');
 
 describe('generateSchema utilities', () => {
   test('isDeviceObject identifies flat objects', () => {
