+++ tests/storage/storage.test.js
@@ -19,7 +19,7 @@ const {
   clearAllData,
   exportAllData,
   importAllData,
-} = require('../storage');
+} = require('../../storage');
 
 const DEVICE_KEY = 'cameraPowerPlanner_devices';
 const SETUP_KEY = 'cameraPowerPlanner_setups';
