+++ tests/ui/sharedLinkButton.test.js
@@ -4,7 +4,7 @@ const { TextEncoder: UtilTextEncoder, TextDecoder: UtilTextDecoder } = require('
 global.TextEncoder = global.TextEncoder || UtilTextEncoder;
 global.TextDecoder = global.TextDecoder || UtilTextDecoder;
 const { JSDOM } = require('jsdom');
-const { getHtmlBody } = require('./domUtils');
+const { getHtmlBody } = require('../domUtils');
 
 test('load shared project button aligns with input', () => {
   const dom = new JSDOM(`<!doctype html><html><head></head><body>${getHtmlBody()}</body></html>`, {
@@ -12,7 +12,7 @@ test('load shared project button aligns with input', () => {
   });
   const { window } = dom;
   const { document } = window;
-  const stylePath = path.join(__dirname, '../style.css');
+  const stylePath = path.join(__dirname, '../../style.css');
   let styleContent = fs.readFileSync(stylePath, 'utf8');
   styleContent = styleContent.replace(/@import[^;]+;/, '');
   const styleEl = document.createElement('style');
